desgin : 추가 정보 등록 모달 UI 수정
